Encapsulated Open/Close. Removed destroy(). Fixed serious bug in directory listings.
o Changed it so that streams are open and closed in the BRStreamInfo
object. Encapsulation is good.
o Got rid of the destroy() method. It made little sense to close files
in super class methods.
o Fixed a major bug in the directory listing method. Original code did
not hold on to bytes that were not
   parsed. As a consequence the directory listings would get horribly
messed up (usually around 18th item).
   Fixed it by reading in the entire directory, then parsing it.