Added cancelRequest, made open timeouts variable.
o Removed some more unused code/code that served no purpose.
o Added timeout to requests. Previously there was a fixed 30 second
timeout for open's. That may be too long for some people.
o Added the ability to cancel uploads and downloads. Canceling an open
at the moment is problematic because of the use of GCD. Need to
evaluate options.
o Added comments to functions