Fixes issue with iOS 7
iOS 6 has a bug that leaks memory when listing a directory. In order to
prevent this from leaking I implemented a "hack" knowing that at some
point Apple would fix this leak.

With iOS 7 fixed that bug which now caused BR to leak OVER release
causing a crash. This fixes that issue.